+++ demo/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@stoplight/elements": "7.0.0-beta.2",
-    "@stoplight/mosaic": "1.0.0-beta.47",
+    "@stoplight/mosaic": "1.0.0-beta.51",
     "history": "^5.0.0",
     "react": "16.14.0",
     "react-dom": "16.14.0",
