+++ packages/elements-core/package.json
@@ -54,7 +54,7 @@
     "@stoplight/json-schema-viewer": "^4.0.0-beta.16",
     "@stoplight/markdown": "^2.10.0",
     "@stoplight/markdown-viewer": "^4.3.3",
-    "@stoplight/mosaic": "1.0.0-beta.47",
+    "@stoplight/mosaic": "1.0.0-beta.51",
     "@stoplight/mosaic-code-editor": "1.0.0-beta.47",
     "@stoplight/mosaic-code-viewer": "1.0.0-beta.47",
     "@stoplight/path": "^1.3.2",
