+++ packages/elements-dev-portal/package.json
@@ -47,7 +47,7 @@
   },
   "dependencies": {
     "@stoplight/elements": "^7.0.0-beta.1",
-    "@stoplight/mosaic": "1.0.0-beta.47",
+    "@stoplight/mosaic": "1.0.0-beta.51",
     "@stoplight/elements-core": "7.0.0-beta.3",
     "lodash": "^4.17.19",
     "react-router-dom": "^5.2.0",
