+++ packages/elements-dev-portal/src/components/Search/index.tsx
@@ -1,26 +1,23 @@
-import { NodeIconMapping } from '@stoplight/elements/types';
-import { Box, Flex, Icon, Input, ListBox, ListBoxItem, Modal, ModalProps, TextColorVals } from '@stoplight/mosaic';
+import { NodeTypeColors, NodeTypeIconDefs } from '@stoplight/elements-core/constants';
+import { Box, Flex, Icon, Input, ListBox, ListBoxItem, Modal, ModalProps } from '@stoplight/mosaic';
 import * as React from 'react';
 
-import { NodeSearchResult } from '../../interfaces/node';
+import { NodeSearchResult } from '../../types';
 
-export const Search = ({
-  search,
-  searchResults,
-  isOpen,
-  onClose,
-  onClick,
-  onSearch,
-}: {
+export type SearchProps = {
   search?: string;
   searchResults?: NodeSearchResult[];
   onSearch: (search: string) => void;
   onClick: (result: NodeSearchResult) => void;
   isOpen?: boolean;
   onClose: ModalProps['onClose'];
-}) => {
-  const listBoxRef = React.useRef<HTMLUListElement>();
+};
+
+export const Search = ({ search, searchResults, isOpen, onClose, onClick, onSearch }: SearchProps) => {
+  const listBoxRef = React.useRef<HTMLDivElement>(null);
+
   const onChange = React.useCallback(e => onSearch(e.currentTarget.value), [onSearch]);
+
   const onKeyDown = React.useCallback(e => {
     // Focus the search results on arrow down
     if (e.key === 'ArrowDown') {
@@ -28,6 +25,7 @@ export const Search = ({
       listBoxRef.current?.focus();
     }
   }, []);
+
   const onSelectionChange = React.useCallback(
     keys => {
       const selectedId = keys.values().next().value;
@@ -48,7 +46,7 @@ export const Search = ({
           appearance="minimal"
           borderB
           size="lg"
-          icon={<Icon icon={['fal', 'search']} size="lg" />}
+          icon="search"
           autoFocus
           placeholder="Search..."
           value={search}
@@ -60,74 +58,59 @@ export const Search = ({
       onClose={onClose}
     >
       {searchResults && searchResults.length > 0 ? (
-        <Box
-          as={ListBox}
+        <ListBox
           ref={listBoxRef}
-          overflowY="auto"
-          style={{ height: '300px' }}
-          m={-5}
           aria-label="Search"
+          overflowY="auto"
+          h={80}
+          m={-5}
           items={searchResults}
           selectionMode="single"
           onSelectionChange={onSelectionChange}
         >
-          {(searchResult: NodeSearchResult) => (
-            <ListBoxItem key={`${searchResult.id}-${searchResult.project_id}`} textValue={searchResult.title}>
-              <Box p={3} borderB>
-                <Flex align="center">
-                  <Box
-                    as={Icon}
-                    size="lg"
-                    w={4}
-                    icon={['fal', NodeIcons[searchResult.type]]}
-                    color={NodeIconColor[searchResult.type]}
-                  />
+          {(searchResult: NodeSearchResult) => {
+            return (
+              <ListBoxItem key={`${searchResult.id}-${searchResult.project_id}`} textValue={searchResult.title}>
+                <Box p={3} borderB>
+                  <Flex align="center">
+                    <Box
+                      as={Icon}
+                      w={4}
+                      icon={NodeTypeIconDefs[searchResult.type]}
+                      style={{ color: NodeTypeColors[searchResult.type] }}
+                    />
+
+                    <Box
+                      flex={1}
+                      fontSize="lg"
+                      dangerouslySetInnerHTML={{ __html: searchResult.highlighted.name }}
+                      fontWeight="medium"
+                      textOverflow="overflow-ellipsis"
+                      mx={2}
+                    />
+
+                    <Box fontSize="sm" color="muted">
+                      {searchResult.project_name}
+                    </Box>
+                  </Flex>
 
                   <Box
-                    flex={1}
-                    fontSize="lg"
-                    dangerouslySetInnerHTML={{ __html: searchResult.highlighted.name }}
-                    fontWeight="medium"
-                    textOverflow="overflow-ellipsis"
-                    mx={2}
+                    dangerouslySetInnerHTML={{ __html: searchResult.highlighted.summary }}
+                    color="muted"
+                    fontSize="sm"
+                    mt={1}
+                    ml={6}
                   />
-
-                  <Box fontSize="sm" color="muted">
-                    {searchResult.project_name}
-                  </Box>
-                </Flex>
-
-                <Box
-                  dangerouslySetInnerHTML={{ __html: searchResult.highlighted.summary }}
-                  color="muted"
-                  fontSize="sm"
-                  mt={1}
-                  ml={6}
-                />
-              </Box>
-            </ListBoxItem>
-          )}
-        </Box>
+                </Box>
+              </ListBoxItem>
+            );
+          }}
+        </ListBox>
       ) : (
-        <Flex w="full" h="full" align="center" justify="center" style={{ height: '300px' }} m={-5}>
+        <Flex w="full" h={80} align="center" justify="center" m={-5}>
           No search results
         </Flex>
       )}
     </Modal>
   );
-};
-
-// TODO: Make these constants in Elements
-const NodeIcons: NodeIconMapping = {
-  http_operation: 'crosshairs',
-  http_service: 'cloud',
-  article: 'book-open',
-  model: 'cube',
-};
-
-const NodeIconColor: Record<string, TextColorVals> = {
-  http_operation: 'success',
-  http_service: 'danger',
-  article: 'primary',
-  model: 'warning',
 };
