+++ packages/elements-dev-portal/src/handlers/getNodes.ts
@@ -1,31 +1,29 @@
-import { NodeSearchResult } from '../interfaces/node';
+import { NodeSearchResult } from '../types';
 
-export const getNodes = async (
-  {
-    workspaceId,
-    projectIds,
-    search,
-    platformUrl = 'https://stoplight.io',
-  }: {
-    workspaceId: string;
-    search?: string;
-    projectIds?: string[];
-    platformUrl?: string;
-  },
-  requestHeaders?: Record<string, string>,
-): Promise<NodeSearchResult[]> => {
+export const getNodes = async ({
+  workspaceId,
+  projectIds,
+  search,
+  platformUrl = 'https://stoplight.io',
+}: {
+  workspaceId: string;
+  search?: string;
+  projectIds?: string[];
+  platformUrl?: string;
+}): Promise<NodeSearchResult[]> => {
   const queryParams = [];
+
   if (projectIds && projectIds.length) {
-    queryParams.push(...projectIds.map(projectId => `project_ids=${projectId}`));
+    queryParams.push(...projectIds.map((projectId, index) => `project_ids[${index}]=${projectId}`));
   }
+
   if (search) {
     queryParams.push(`search=${search}`);
   }
+
   const query = queryParams.length ? `?${queryParams.join('&')}` : '';
 
-  const response = await fetch(`${platformUrl}/api/v1/workspaces/${workspaceId}/nodes${query}`, {
-    headers: requestHeaders,
-  });
+  const response = await fetch(`${platformUrl}/api/v1/workspaces/${workspaceId}/nodes${query}`);
   const data = await response.json();
 
   if (!response.ok) {
