+++ packages/elements-dev-portal/src/index.ts
@@ -2,15 +2,19 @@ export type { BranchSelectorProps } from './components/BranchSelector';
 export { BranchSelector } from './components/BranchSelector';
 export type { NodeContentProps } from './components/NodeContent';
 export { NodeContent } from './components/NodeContent';
+export type { SearchProps } from './components/Search';
+export { Search } from './components/Search';
 export type { TableOfContentsProps } from './components/TableOfContents';
 export { TableOfContents } from './components/TableOfContents';
 export type { StoplightProjectProps } from './containers/StoplightProject';
 export { StoplightProject } from './containers/StoplightProject';
 export { getBranches } from './handlers/getBranches';
 export { getNodeContent } from './handlers/getNodeContent';
+export { getNodes } from './handlers/getNodes';
 export { getTableOfContents } from './handlers/getTableOfContents';
 export { useGetBranches } from './hooks/useGetBranches';
 export { useGetNodeContent } from './hooks/useGetNodeContent';
+export { useGetNodes } from './hooks/useGetNodes';
 export { useGetTableOfContents } from './hooks/useGetTableOfContents';
 export type { ProjectTableOfContents } from './types';
 export type { Branch, Node } from './types';
